+++ app/src/main/java/io/zoemeow/dutschedule/service/NewsUpdateService.kt
@@ -1,7 +1,7 @@
 package io.zoemeow.dutschedule.service
 
 import android.content.Intent
-import io.dutwrapperlib.dutwrapper.model.enums.LessonStatus
+import io.dutwrapper.dutwrapper.model.enums.LessonStatus
 import io.zoemeow.dutschedule.activity.PermissionRequestActivity
 import io.zoemeow.dutschedule.model.ProcessState
 import io.zoemeow.dutschedule.model.permissionrequest.PermissionList